Agent-based model – копія/requirements.txt
# requirements.txt
mesa>=3.5.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.0
